llm-server/pyproject.toml
[project]
name = "llm-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.13.0",
    "fastapi>=0.135.3",
    "huggingface-hub>=1.9.0",
    "llama-cpp-python>=0.3.20",
    "pydantic>=2.12.5",
    "torch>=2.11.0",
    "transformers>=5.5.0",
    "uvicorn>=0.44.0",
]
